requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
google-api-python-client==2.108.0
python-dotenv==1.0.0
markdown==3.5.1
Flask-Markdown==0.3 